feat(progress-bar): add linear progress bars and battery widget support
- Refactor `circular_progress_bar.py` into a unified `progress_bar.py`
- Add `progress_type` config to support `circular`, `linear_horizontal`, and `linear_vertical` styles
- Add configurable `radius` for linear progress bars
- Inject the unified progress bar widget across all supporting widgets
- Implement progress bar support for the Battery widget (closes #945)
- BREAKING CHANGE: The CSS class for progress bars has been updated from `.progress-circle` to `.progress-container` across all widgets
