+++ src/core/utils/utilities.py
@@ -62,21 +62,23 @@ def refresh_widget_style(*widgets: QWidget) -> None:
 
 
 def build_progress_widget(self, options: dict[str, Any]) -> None:
-    """Builds a circular progress widget based on the provided options."""
+    """Builds a circular or linear progress widget based on the provided options."""
     if not options["enabled"]:
         return
 
-    from core.utils.circular_progress_bar import CircularProgressBar, CircularProgressWidget
+    from core.utils.progress_bar import ProgressBar, ProgressWidget
 
-    self.progress_data = CircularProgressBar(
+    self.progress_data = ProgressBar(
         parent=self,
         size=options["size"],
         thickness=options["thickness"],
         color=options["color"],
         background_color=options["background_color"],
         animation=options["animation"],
+        progress_type=options.get("progress_type", "circular"),
+        radius=options.get("radius", 0),
     )
-    self.progress_widget = CircularProgressWidget(self.progress_data)
+    self.progress_widget = ProgressWidget(self.progress_data)
     return self.progress_widget
