+++ src/core/validation/widgets/yasb/battery.py
@@ -1,3 +1,5 @@
+from typing import Literal
+
 from pydantic import Field
 
 from core.validation.widgets.base_model import (
@@ -5,6 +7,18 @@ from core.validation.widgets.base_model import (
     CustomBaseModel,
     KeybindingConfig,
 )
+
+
+class ProgressBarConfig(CustomBaseModel):
+    enabled: bool = False
+    progress_type: Literal["circular", "linear_horizontal", "linear_vertical"] = "circular"
+    size: int = Field(default=18, ge=1, le=200)
+    thickness: int = Field(default=3, ge=1, le=100)
+    radius: int = Field(default=0, ge=0, le=100)
+    color: str | list[str] = "#00C800"
+    background_color: str = "#3C3C3C"
+    position: Literal["left", "right"] = "left"
+    animation: bool = True
 
 
 class ChargingOptionsConfig(CustomBaseModel):
@@ -45,5 +59,6 @@ class BatteryConfig(CustomBaseModel):
     charging_options: ChargingOptionsConfig = ChargingOptionsConfig()
     status_thresholds: StatusThresholdsConfig = StatusThresholdsConfig()
     status_icons: StatusIconsConfig = StatusIconsConfig()
+    progress_bar: ProgressBarConfig = ProgressBarConfig()
     keybindings: list[KeybindingConfig] = []
     callbacks: CallbacksBatteryConfig = CallbacksBatteryConfig()
